requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
configparser==5.1.0
datasets==1.15.1
debugpy==1.5.1
decorator==5.1.0
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.3
faiss-cpu==1.7.1.post2
filelock==3.4.0
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
higher==0.2.1
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==1.5.0
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
Jinja2==3.0.3
joblib==0.15.0
jupyter-client==7.0.6
jupyter-core==4.9.1
lmdb==0.98
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
-e git+https://github.com/yuchenlin/MetaCross.git@e139e4f412b8a9300e5f97aadbab6b57ce5733c5#egg=metax
methodtools==0.1.2
multidict==5.2.0
multiprocess==0.70.12.2
multivolumefile==0.2.3
nest-asyncio==1.5.1
nltk==3.5
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1636145336524/work
packaging==21.3
pandas==1.0.3
parso==0.8.2
path==13.1.0
path.py==12.4.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.22
-e git+https://github.com/bigscience-workshop/promptsource.git@b1736edcbaca4bfd997c6f068166113299e3fb97#egg=promptsource
protobuf==3.12.1
psutil==5.8.0
ptyprocess==0.7.0
py-lz4framed==0.14.0
py-rouge==1.1
py7zr==0.16.3
pyarrow==6.0.1
pybcj==0.5.0
pycryptodomex==3.11.0
Pygments==2.10.0
pyparsing==3.0.6
pyppmd==0.17.3
pyrouge==0.1.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.0
regex==2020.5.14
requests==2.26.0
sacremoses==0.0.46
scikit-learn==0.24.1
scipy==1.4.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
sentry-sdk==1.5.0
shortuuid==1.0.8
six==1.15.0
smmap==5.0.0
subprocess32==3.5.4
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==2.0.0
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.5
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
urllib3==1.26.7
wandb==0.12.7
wcwidth==0.2.5
wget==3.2
wirerope==0.3.1
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.0.0
